Split the route drawing to its own page
